+++ index.js
@@ -3,14 +3,14 @@ const postcss = require('postcss');
 
 // plugins by name
 const plugins = {
-	border: 'postcss-short-border',
-	borderRadius: 'postcss-short-border-radius',
-	color: 'postcss-short-color',
-	fontSize: 'postcss-short-font-size',
-	position: 'postcss-short-position',
-	size: 'postcss-short-size',
-	spacing: 'postcss-short-spacing',
-	fontWeights: 'postcss-font-weights'
+	border:       require('postcss-short-border'),
+	borderRadius: require('postcss-short-border-radius'),
+	color:        require('postcss-short-color'),
+	fontSize:     require('postcss-short-font-size'),
+	position:     require('postcss-short-position'),
+	size:         require('postcss-short-size'),
+	spacing:      require('postcss-short-spacing'),
+	fontWeights:  require('postcss-font-weights')
 };
 
 // plugin
@@ -18,20 +18,30 @@ module.exports = postcss.plugin('postcss-short', (opts = {}) => {
 	// cached processor
 	const processor = postcss();
 
-	Object.keys(plugins).forEach((name) => {
-		// options by name
-		const pluginOpts = Object.assign({
-			disable: opts[name] === false
-		}, opts[name]);
+	// for each plugin
+	Object.keys(plugins).forEach(
+		(name) => {
+			// plugin options by name
+			const pluginOpts = Object.assign({
+				disable: opts[name] === false
+			}, opts[name]);
 
-		if (!pluginOpts.disable) {
-			// cached plugin
-			const plugin = plugins[name] = typeof plugins[name] === 'string' ? require(plugins[name]) : plugins[name];
-
-			// use the plugin with the processor
-			processor.use(plugin(pluginOpts));
+			// if the plugin is not disabled
+			if (!pluginOpts.disable) {
+				// use the plugin
+				processor.use(
+					plugins[name](pluginOpts)
+				);
+			}
 		}
-	});
+	);
 
 	return processor;
 });
+
+// override plugin#process
+module.exports.process = function (cssString, pluginOptions, processOptions) {
+	return postcss([
+		0 in arguments ? module.exports(pluginOptions) : module.exports()
+	]).process(cssString, processOptions);
+};